Fixes several issues, including issue #13 on Github.
Note that the Github issue is one I encountered as well. The problem
appears to be caused by which-key, and thus, may not be a treeclimber
issue at all, but because which-key is so ubiquitous, it probably makes
sense to put a small workaround in treeclimber, pending a more complete
fix.

In addition to that fatal error, there were a number of maps that just
didn't work correctly. A not necessarily exhaustive list is presented
below...

select_grow_forward/backward were broken by the presence of co-located
nodes (i.e., distinct nodes with the same extents, one a wrapper for the
other). The problem was that largest_named_descendant_for_range()
selected the outermost co-located node, with the result that the
children of the innermost weren't considered as candidates by
get_covering_nodes().

select_shrink was broken because of issues with the history stack. For
one thing, nodes were pushed onto the stack by select_expand, but *not*
by select_current_node. Thus, for a sequence like....
	select_current_node
	select_expand
	select_shrink
...the originally selected node was not in the stack, and thus was not
considered as a possible target for the shrink. Even worse, target nodes
were not required to be below the current node, which meant that the
shrink selected the *current* node in this case, making it effectively a
no-op.

TODO: On another commit, need to rework history stack logic a bit. I'm
thinking it makes sense to push only on ascent. Moreover, I'd like to
make the following statement from the README (specifically the "or"
clause) true: "Shrinking selects a child node from the current node, or
will undo the action of a previous expand operation." Currently,
shrinking doesn't really undo, as the shrink logic will never select
more than one child node, even when the expand was performed with
a *range* of child nodes selected.