Refactored the node history stack to fix select_shrink.
Originally, nodes were pushed to the stack whenever nodes were visited,
without regard to the stack's current contents or the type of movement
(ascent, descent or lateral). This led to unexpected behavior from
select_shrink, which, whenever possible, relied on the stack to find the
shrink target.

Another problem with the original approach is that the docs state that a
shrink should undo a previous expand, but this was not the case when
multiple nodes were selected prior to the expand. Interestingly,
*ranges*, not nodes, were being pushed to the stack, but the logic
forced selection of a single node as the shrink target.

In the new approach, we push to the stack only on ascent, and when
pulling from the stack, translate a multi-node range to a multi-node
shrink target. The stack is cleared whenever we descend through a child
that's outside the range at the top of the stack.
Note: Selecting the top level node pushes all the ascended-through
nodes, including the initial range/node.