Fixed issue with apply_decoration.
Sibling nodes always used the current node's start point, which broke
highlighting in the multi-line case.
Note: Now that highlighting attributes like bold and italic are
supported, the function needs to be reworked to avoid creating a single,
all-encompassing "parent" group. This works for bg colors, but the
highighting attributes are applied to all nested groups. Could introduce
an option governing which method is used (in case user wants the
inherited attributes), but I think the default should be no inheritance.