Made the default <M-k> mapping in normal mode refer to select_current_node.
The old default had <M-k> mapped to select_expand in n, x, and o modes.
This worked because select_current_node and select_expand are
functionally equivalent in normal mode; however, not having a map for
select_current_node in normal mode could be a source of confusion to the
user.

Removed bold from default Selection decoration.